fix: final-check

+++ LibraryManagementSystem.Application/Interfaces/IBookService.cs
@@ -5,9 +5,9 @@ namespace LibraryManagementSystem.Application.Interfaces
     public interface IBookService
     {
         Task<ApiResponse> CreateAsync(BookDtos.CreateRequest dto);
-        Task<ApiResponse> DeleteAsync(long id);
+        Task<ApiResponse> DeleteAsync(int id);
         Task<ApiResponse> GetAllAsync(BookDtos.BookFilterRequest request);
-        Task<ApiResponse> GetByIdAsync(long id);
-        Task<ApiResponse> UpdateAsync(long id, BookDtos.UpdateRequest dto);
+        Task<ApiResponse> GetByIdAsync(int id);
+        Task<ApiResponse> UpdateAsync(int id, BookDtos.UpdateRequest dto);
     }
 }
